+++ Changelog.md
@@ -1,6 +1,13 @@
+Version 1.1 (?? October 2015)
+-----------------------------
+
+* Same as cf/units.py in cf-python version 1.1
+
+* Added better version checking in __init__.py
+	
 Version 1.0 (27 May 2015)
 -------------------------
 
 * Initial release
 
-* Same as cf/units.py in cf-python
+* Same as cf/units.py in cf-python version 1.0